fix(frontend): count compatibility occurrences in banner, share grouping logic
