+++ src/components/dashboard/CompatibilityBanner.vue
@@ -3,6 +3,7 @@ import { ref, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useRouter } from 'vue-router'
 import IconAlertTriangle from '~icons/lucide/alert-triangle'
+import { groupCompatibilityEvents } from '~/services/compatibilityEvents'
 import { useSupabase } from '~/services/supabase'
 
 const props = defineProps<{
@@ -22,9 +23,12 @@ async function fetchUnresolvedCount() {
   }
 
   try {
-    const { count, error } = await supabase
+    // Count occurrences, not raw rows: one channel change is many per-platform
+    // rows. Group the unresolved rows the same way the history page does so the
+    // banner count matches what the user sees there.
+    const { data, error } = await supabase
       .from('compatibility_events')
-      .select('*', { count: 'exact', head: true })
+      .select('id, platform, channel_id, current_version_id, previous_version_id, source, change_occurred_at, created_at, resolved_at')
       .eq('app_id', props.appId)
       .is('resolved_at', null)
 
@@ -34,7 +38,7 @@ async function fetchUnresolvedCount() {
       return
     }
 
-    unresolvedCount.value = count ?? 0
+    unresolvedCount.value = groupCompatibilityEvents(data ?? []).length
   }
   catch (error) {
     console.error('[CompatibilityBanner] Error fetching unresolved count:', error)
